+++ nens_auth_client/tests/test_authorize_view.py
@@ -306,7 +306,7 @@ def test_authorize_wrong_state(id_token_generator, auth_req_generator):
     # we solve this by restarting the login process (keeping the success_url).
     id_token, claims = id_token_generator()
     request = auth_req_generator(id_token, state="a")
-    request.session.pop("_state_cognito_a")
+    request.session.pop("_state_oauth_a")
     request.session[LOGIN_REDIRECT_SESSION_KEY] = "/success"
     response = views.authorize(request)
 
@@ -400,7 +400,7 @@ def test_token_error(rq_mocker, rf, openid_configuration):
     )
     # Create the request
     request = rf.get("http://testserver/authorize/?code=abc&state=my_state")
-    request.session = {"_state_cognito_my_state": {"data": {"nonce": "x"}}}
+    request.session = {"_state_oauth_my_state": {"data": {"nonce": "x"}}}
     with pytest.raises(OAuthError, match="some_error: bla"):
         views.authorize(request)
 
@@ -416,7 +416,7 @@ def test_token_error_code_already_used(rq_mocker, rf, openid_configuration):
     )
     # Create the request
     request = rf.get("http://testserver/authorize/?code=abc&state=my_state")
-    request.session = {"_state_cognito_my_state": {"data": {"nonce": "bla"}}}
+    request.session = {"_state_oauth_my_state": {"data": {"nonce": "bla"}}}
     response = views.authorize(request)
 
     assert response.status_code == 302