+++ nens_auth_client/wso2.py
@@ -1,6 +1,4 @@
 from authlib.integrations.django_client import DjangoOAuth2App
-from authlib.jose import JsonWebKey
-from authlib.jose import JsonWebToken
 from django.http.response import HttpResponseRedirect
 from urllib.parse import urlencode
 from urllib.parse import urlparse
@@ -45,39 +43,15 @@ class WSO2AuthClient(DjangoOAuth2App):
         return HttpResponseRedirect(logout_url)
 
     def parse_access_token(self, token, claims_options=None, leeway=120):
-        # this is a copy from the _parse_id_token equivalent function
-        def load_key(header, payload):
-            jwk_set = self.fetch_jwk_set()
-            kid = header.get("kid")
-            try:
-                return JsonWebKey.import_key_set(jwk_set).find_by_kid(kid)
-            except ValueError:
-                # re-try with new jwk set
-                jwk_set = self.fetch_jwk_set(force=True)
-                return JsonWebKey.import_key_set(jwk_set).find_by_kid(kid)
+        # TODO:
+        raise NotImplementedError()
 
-        metadata = self.load_server_metadata()
-        claims_options = {
-            "iss": {"essential": True, "value": metadata["issuer"]},
-            "sub": {"essential": True},
-            **(claims_options or {}),
-        }
-
-        alg_values = metadata.get("id_token_signing_alg_values_supported")
-        if not alg_values:
-            alg_values = ["RS256"]
-
-        claims = JsonWebToken(alg_values).decode(
-            token, key=load_key, claims_options=claims_options
-        )
-
-        claims.validate(leeway=leeway)
-        return claims
-
-    def extract_provider_name(self, claims):
+    @staticmethod
+    def extract_provider_name(claims):
         """Return provider name from claim and `None` if not found"""
         return None
 
-    def extract_username(self, claims) -> str:
+    @staticmethod
+    def extract_username(claims) -> str:
         """Return username from claims"""
         return claims["email"]